# src/main/resources/static/docs/openapi-v0.0.4.yaml
openapi: 3.0.1
info:
  title: IITP Open API Documentation
  description: IITP Open API 문서입니다.
  version: v0.0.4
...
servers:
- url: http://<SERVER_IP>/mock
  description: Prism Mock 서버
security:
- ApiKeyAuth: []
paths:
...
  /api/v1/basic/social/social/particFreq/{statYear}:
    get:
      tags:
      - 기초-사회망 현황 API V1
      summary: 해당 연도의 장애인의 사회 참여 조회(연도)
      description: 해당 연도의 장애인의 사회 참여 조회
      operationId: getSocialParticFreqYear
      parameters:
      - name: statYear
        in: path
        description: 통계 연도
        required: true
        schema:
          type: integer
          format: int32
        example: 2024
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResDtoStatDataRes'
  /api/v1/basic/social/social/particFreq/latest:
    get:
      tags:
      - 기초-사회망 현황 API V1
      summary: 장애인의 사회 참여 조회(최신)
      description: 장애인의 사회 참여 조회. (최대 10년) 예) from(옵션)~to(옵션), default는 최종 3개년
      operationId: getSocialParticFreqLatest
      parameters:
      - name: from
        in: query
        description: '(옵션) 통계 조회 시작 연도  (최대: 10년 조회)'
        required: false
        schema:
          type: integer
          format: int32
        example: 2019
      - name: to
        in: query
        description: '(옵션) 통계 조회 끝 연도  (최대: 10년 조회)'
        required: false
        schema:
          type: integer
          format: int32
        example: 2022
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResDtoStatDataRes'
...
  /api/v1/basic/social/social/contractFreq/{statYear}:
    get:
      tags:
      - 기초-사회망 현황 API V1
      summary: 해당 연도의 가까이 지내는 친구, 이웃, 지인 수 및 만남 빈도 조회(연도)
      description: 해당 연도의 가까이 지내는 친구, 이웃, 지인 수 및 만남 빈도 조회
      operationId: getSocialContactCntfreqYear
      parameters:
      - name: statYear
        in: path
        description: 통계 연도
        required: true
        schema:
          type: integer
          format: int32
        example: 2024
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResDtoStatDataRes'
  /api/v1/basic/social/social/contractFreq/latest:
    get:
      tags:
      - 기초-사회망 현황 API V1
      summary: 가까이 지내는 친구, 이웃, 지인 수 및 만남 빈도 조회(최신)
      description: 가까이 지내는 친구, 이웃, 지인 수 및 만남 빈도. (최대 10년) 예) from(옵션)~to(옵션), default는
        최종 3개년
      operationId: getSocialContactCntfreqLatest
      parameters:
      - name: from
        in: query
        description: '(옵션) 통계 조회 시작 연도  (최대: 10년 조회)'
        required: false
        schema:
          type: integer
          format: int32
        example: 2019
      - name: to
        in: query
        description: '(옵션) 통계 조회 끝 연도  (최대: 10년 조회)'
        required: false
        schema:
          type: integer
          format: int32
        example: 2022
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResDtoStatDataRes'
...
components:
  schemas:
...
    ErrInfoDto:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 에러 메시지
          example: 요청한 parameter이 올바르지 않습니다.
        details:
          type: string
          description: 에러 상세 설명
          example: 파라미터 'id'는 필수입니다.
      description: Response 에러 상세 정보
    ErrorCode:
      type: string
      description: 에러 코드
      example: C000
      enum:
      - S000
      - C000
      - C001
      - C002
      - C003
      - C004
      - C005
      - C006
      - C007
      - C008
      - C009
      - C010
      - C011
      - B001
      - B002
      - B003
      - B004
      - S001
      - P001
      - P002
      - P003
      - P004
...
    ApiResDtoStatDataRes:
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
          example: true
        data:
          $ref: '#/components/schemas/StatDataRes'
        status:
          type: string
          description: 응답 코드
          example: '200'
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 EARLY_HINTS
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
        error:
          $ref: '#/components/schemas/ErrInfoDto'
      description: 공통 응답 포맷
    StatDataItem:
      required:
      - c1Nm
      - c1ObjId
      - c1ObjNm
      - data
      - itmNm
      - itmObjId
      - prdDe
      - unitNm
      type: object
      properties:
        prdDe:
          type: integer
          description: 통계 년도
          format: int32
          example: 2021
        c1Nm:
          type: string
          description: 분류값 1
          example: 전체
        c2Nm:
          type: string
          description: 분류값 2
          example: 추정수
        c3Nm:
          type: string
          description: 분류값 3
        c1ObjId:
          type: string
          description: '분류코드 1 '
          example: A
        c1ObjNm:
          type: string
          description: 분류명 1
          example: 특성별
        c2ObjId:
          type: string
          description: '분류코드 2 '
          example: B
        c2ObjNm:
          type: string
          description: 분류명 2
          example: 도와주는사람별
        c3ObjId:
          type: string
          description: '분류코드 3 '
        c3ObjNm:
          type: string
          description: 분류명 3
        itmNm:
          type: string
          description: 항목값
          example: 일상생활 도와주는 사람(1순위)
        itmObjId:
          type: string
          description: 항목코드
          example: ITEM
        unitNm:
          type: string
          description: 단위명
          example: 명
        data:
          type: string
          description: 수치 값
          example: '251277'
        lstChnDe:
          type: string
          description: 데이터의 최종수정일
          format: date
          example: '2023-01-02'
      description: 통계 데이터 상세 Item
    StatDataRes:
      type: object
      properties:
        items:
          type: array
          description: 통계 데이터 리스트
          items:
            $ref: '#/components/schemas/StatDataItem'
      description: 통계 데이터
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: API Key for authentication
      name: X-API-KEY
      in: header